prescription-service/requirements.txt
Django>=4.2
djangorestframework
djangorestframework-simplejwt
django-environ
requests
psycopg2-binary
pika
